% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can-be-num.R
\name{str_can_be_numeric}
\alias{str_can_be_numeric}
\title{Check if a string could be considered as numeric.}
\usage{
str_can_be_numeric(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A logical vector.
}
\description{
After padding is removed, could the input string be considered to be numeric,
i.e. could it be coerced to numeric. This function is vectorized over its one
argument.
}
\examples{
str_can_be_numeric("3")
str_can_be_numeric("5 ")
str_can_be_numeric(c("1a", "abc"))
}
\concept{type converters}
