% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{str_extract_currencies}
\alias{str_extract_currencies}
\title{Extract currency amounts from a string.}
\usage{
str_extract_currencies(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A \link[tibble:tibble-package]{tibble} with 4 columns: \code{string_num},
\code{string}, \code{curr_sym} and \code{amount}. Every extracted currency amount gets its
own row in the tibble detailing the string number and string that it was
extracted from, the currency symbol and the amount.
}
\description{
The currency of a number is defined as the character coming before the number
in the string. If nothing comes before (i.e. if the number is the first thing
in the string), the currency is the empty string, similarly the currency can
be a space, comma or any manner of thing. \code{str_extract_currencies()} takes a
string and returns all of the currency amounts numbers within that string. It
is vectorized over \code{string}.
}
\details{
\itemize{\item "-$2.00" and "$-2.00" are interpreted as negative two dollars.
\item If you request e.g. the 5th currency amount but there are only 3
currency amounts, you get an amount and currency symbol of \code{NA}.}
}
\examples{
str_extract_currencies(c(
  "35.00 $1.14", "abc5 $3.8 77",
  "-$1.5e6", "over £1,000"
))
}
\seealso{
\code{\link[=str_nth_currency]{str_nth_currency()}}
}
