% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{str_locate_braces}
\alias{str_locate_braces}
\title{Locate the braces in a string.}
\usage{
str_locate_braces(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
A list of data frames, one for each member of the string character
vector. Each data frame has a "position" and "brace" column which give the
positions and types of braces in the given string.
}
\description{
Give the positions of \code{(}, \code{)}, \code{[}, \code{]}, \code{\{}, \code{\}} within a string.
}
\examples{
str_locate_braces(c("a{](kkj)})", "ab(]c{}"))
}
