% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{str_locate_nth}
\alias{str_locate_nth}
\alias{str_locate_first}
\alias{str_locate_last}
\title{Locate the indices of the \eqn{n}th instance of a pattern.}
\usage{
str_locate_nth(string, pattern, n)

str_locate_first(string, pattern)

str_locate_last(string, pattern)
}
\arguments{
\item{string}{A character vector. These functions are vectorized over this
argument.}

\item{pattern}{A character vector. Pattern(s) specified like the pattern(s)
in the stringr package (e.g. look at \code{\link[stringr:str_locate]{stringr::str_locate()}}). If
this has length >1 its length must be the same as that of \code{string}.}

\item{n}{A character vector of length 1 or the same length as \code{string}.}
}
\value{
A two-column matrix. The \eqn{i}th row of this matrix gives the start
and end indices of the \eqn{n}th instance of \code{pattern} in the \eqn{i}th
element of \code{string}.
}
\description{
The \eqn{n}th instance of an pattern will cover a series of character
indices. These functions tell you which indices those are. These functions
are vectorised over all arguments.
}
\details{
\itemize{ \item \code{str_locate_first(...)} is just \code{str_locate_nth(..., n = 1)}.
\item \code{str_locate_last(...)} is just \code{str_locate_nth(..., n = -1)}. }
}
\examples{
str_locate_nth(c("abcdabcxyz", "abcabc"), "abc", 2)
str_locate_nth(c("This old thing.", "That beautiful thing there."),
"\\\\w+", c(2, -2))
str_locate_nth("abc", "b", c(0, 1, 1, 2))

}
\seealso{
Other locators: \code{\link{str_locate_braces}}
}
\concept{locators}
