% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{str_nth_currency}
\alias{str_nth_currency}
\alias{str_first_currency}
\alias{str_last_currency}
\title{Extract the \code{n}th currency amount from a string.}
\usage{
str_nth_currency(string, n)

str_first_currency(string)

str_last_currency(string)
}
\arguments{
\item{string}{A character vector.}

\item{n}{A numeric vector of length 1 or the same length as \code{string}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the same style as the return
from \code{\link[=str_extract_currencies]{str_extract_currencies()}}. 4 columns: \code{string_num}, \code{string},
\code{curr_sym} and \code{amount}. Every extracted currency amount gets its own row
in the tibble detailing the string number and string that it was extracted
from, the currency symbol and the amount.
}
\description{
These functions are vectorized over \code{string} and \code{n}.
\code{\link[=str_extract_currencies]{str_extract_currencies()}} extracts all currency amounts but
\code{str_nth_currency()} just gets the \code{n}th currency amount from each string.
\code{str_first_currency(string)} and \code{str_last_currency(string)} are just
wrappers for \code{str_nth_currency(string, n = 1)} and \code{str_nth_currency(string, n = -1)}.
}
\details{
\itemize{\item "-$2.00" and "$-2.00" are interpreted as negative two dollars.
\item If you request e.g. the 5th currency amount but there are only 3
currency amounts, you get an amount and currency symbol of \code{NA}.}
}
\examples{
string <- c("ab3 13", "$1", "35.00 $1.14", "abc5 $3.8", "stuff")
str_nth_currency(string, n = 2)
str_nth_currency(string, c(1, 2, 1, 2, 1))
str_first_currency(string)
str_last_currency(string)
}
