% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem.R
\name{str_paste_elems}
\alias{str_paste_elems}
\title{Extract bits of a string and paste them together}
\usage{
str_paste_elems(string, indices)
}
\arguments{
\item{string}{A string.}

\item{indices}{A numeric vector of positive integers detailing the indices of
the characters of \code{string} that we wish to paste together.}
}
\value{
A string.
}
\description{
Extract characters - specified by their indices - from a string and paste
them together
}
\examples{
str_paste_elems("abcdef", c(2, 5:6))
}
