% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{str_remove_quoted}
\alias{str_remove_quoted}
\title{Remove the quoted parts of a string.}
\usage{
str_remove_quoted(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector.
}
\description{
If any parts of a string are quoted (between quotation marks), remove those
parts of the string, including the quotes. Run the examples and you'll know
exactly how this function works.
}
\examples{
string <- "\\"abc\\"67a\\'dk\\'f"
cat(string)
str_remove_quoted(string)

}
\seealso{
Other removers: \code{\link{str_singleize}},
  \code{\link{str_trim_anything}}
}
\concept{removers}
