% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel-case.R
\name{str_split_camel_case}
\alias{str_split_camel_case}
\title{Split a string based on CamelCase.}
\usage{
str_split_camel_case(string, lower = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{lower}{Do you want the output to be all lower case (or as is)?}
}
\value{
A list of character vectors, one list element for each element of
\code{string}.
}
\description{
Vectorized over \code{string}.
}
\examples{
str_split_camel_case(c("RoryNolan", "NaomiFlagg", "DepartmentOfSillyHats"))
str_split_camel_case(c("RoryNolan", "NaomiFlagg", "DepartmentOfSillyHats",
  lower = TRUE
))
}
\references{
Adapted from Ramnath Vaidyanathan's answer at
http://stackoverflow.com/questions/8406974/splitting-camelcase-in-r.
}
\seealso{
Other splitters: \code{\link{str_split_by_numbers}}
}
\concept{splitters}
