% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strex-package.R
\docType{package}
\name{strex}
\alias{strex}
\alias{strex-package}
\title{\code{strex}: extra string manipulation functions}
\description{
There are some things that I wish were easier with the \code{stringr} or \code{stringi}
packages. The foremost of these is the extraction of numbers from strings.
\code{stringr} makes you figure out the regex for yourself; \code{strex} takes care of
this for you. There are many more useful functionalities in \code{strex}. In
particular, there's a \code{match_arg()} function which is more flexible than the
base \code{match.arg()}. Contributions to this package are encouraged: it is
intended as a miscellany of string manipulation functions which cannot be
found in \code{stringi} or \code{stringr}.
}
\references{
Rory Nolan and Sergi Padilla-Parra (2017). filesstrings: An R
package for file and string manipulation. The Journal of Open Source
Software, 2(14).  \doi{10.21105/joss.00260}.
}
